backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
SQLAlchemy>=2.0.30
alembic>=1.13
pydantic>=2.6
python-dotenv>=1.0
httpx>=0.27
requests>=2.31
opencv-python-headless>=4.8
pytesseract>=0.3
rapidfuzz>=3.5
langchain>=0.2
langchain-google-genai>=1.0.0
google-generativeai>=0.6.0
sentence-transformers>=2.2
faiss-cpu>=1.7
numpy>=1.26
gTTS>=2.4
python-multipart>=0.0.9
uvloop>=0.19 ; platform_system != "Windows"
loguru>=0.7
pytest>=8
pytest-asyncio>=0.23
starlette>=0.37
typing-extensions>=4.9
